Delete files covered by foundry-vtt-react-application
