+++ src/applications/my-react-component.jsx
@@ -1,0 +1,3 @@
+export default function MyReactComponent(props) {
+  return <div>Hello, {props.data}! Stinky ass BOUETCH. Sorry OK.</div>;
+}
