+++ src/applications/reactor-sheet.js
@@ -1,7 +1,8 @@
 import ReactorSheetApp from "@src/ReactorSheet";
-import ReactApplication from "./react-application";
 
-class ReactorSheet extends ReactApplication {
+import { ReactActorSheetV2 } from "foundry-vtt-react-application";
+
+class ReactorSheet extends ReactActorSheetV2 {
   reactApp = ReactorSheetApp;
   static DEFAULT_OPTIONS = {
     window: {
@@ -34,15 +35,11 @@ class ReactorSheet extends ReactApplication {
 
   static async #onEditImage(e, target) {
     if (target.nodeName !== "IMG") {
-      throw new Error(
-        "The editImage action is available only for IMG elements."
-      );
+      throw new Error("The editImage action is available only for IMG elements.");
     }
     const attr = target.dataset.edit;
     const current = foundry.utils.getProperty(this.document._source, attr);
-    const defaultArtwork =
-      this.document.constructor.getDefaultArtwork?.(this.document._source) ??
-      {};
+    const defaultArtwork = this.document.constructor.getDefaultArtwork?.(this.document._source) ?? {};
     const defaultImage = foundry.utils.getProperty(defaultArtwork, attr);
     const fp = new foundry.applications.apps.FilePicker.implementation({
       current,
